Move pkg/http-handler to kubeops
